Provide a fast-path for suitable argument types.
When implementing specializations always think of what argument types could potentially occur, but also which are most likely to occur. A more generic specialization using InteropLibrary to coerce argument values can be a necessary fallback, but the additional node children make the AST more complex and unnecessarily inefficient when the most likely argument types are encountered. We can expect that this node will mostly be executed with arguments that are already int, so an additional specialization makes sense here. As long as this common case occurs the Truffle DSL can then avoid creating the libraries in favor of just doing a more efficient inlined type check.